+++ tests/client/test_secrets.py
@@ -52,7 +52,7 @@ def test_secret_value_depends_on_use_local_secrets(monkeypatch):
 
 
 def test_secrets_use_client(monkeypatch):
-    response = {"data": {"secretValue": "1234"}}
+    response = {"data": {"secretValue": '"1234"'}}
     post = MagicMock(return_value=MagicMock(json=MagicMock(return_value=response)))
     session = MagicMock()
     session.return_value.post = post
@@ -82,6 +82,21 @@ def test_cloud_secrets_remain_plain_dictionaries(monkeypatch):
     assert isinstance(val2, list) and not isinstance(val2, box.BoxList)
     val3 = val["b"][2]
     assert isinstance(val3, dict) and not isinstance(val3, box.Box)
+
+
+def test_cloud_secrets_auto_load_json_strings(monkeypatch):
+    response = {"data": {"secretValue": '{"x": 42}'}}
+    post = MagicMock(return_value=MagicMock(json=MagicMock(return_value=response)))
+    session = MagicMock()
+    session.return_value.post = post
+    monkeypatch.setattr("requests.Session", session)
+    with set_temporary_config(
+        {"cloud.auth_token": "secret_token", "cloud.use_local_secrets": False}
+    ):
+        my_secret = Secret(name="the-key")
+        val = my_secret.get()
+
+    assert isinstance(val, dict)
 
 
 def test_local_secrets_auto_load_json_strings():